requirements.txt
absl-py==2.0.0
apturl==0.5.2
asgiref==3.5.0
astunparse==1.6.3
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
cachetools==5.3.1
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
contourpy==1.1.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==3.2.12
django-cors-headers==4.2.0
djangorestframework==3.14.0
djongo==1.3.6
duplicity==0.8.21
et-xmlfile==1.1.0
fasteners==0.14.1
filelock==3.12.3
flatbuffers==23.5.26
fonttools==4.43.0
future==0.18.2
gast==0.5.4
google-auth==2.23.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
gunicorn==21.2.0
h5py==3.9.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
joblib==1.3.2
keras==2.14.0
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==16.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.8.0
ml-dtypes==0.2.0
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.26.0
oauthlib==3.2.0
olefile==0.46
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.1.1
paramiko==2.9.3
pbr==5.11.1
pexpect==4.8.0
Pillow==9.0.1
platformdirs==3.10.0
protobuf==4.24.3
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.20.1
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==3.5.1
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.1
scipy==1.11.3
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.2.4
stevedore==5.1.0
systemd-python==234
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
threadpoolctl==3.2.0
typing_extensions==4.7.1
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.24.5
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wadllib==1.3.6
Werkzeug==3.0.0
wrapt==1.14.1
xdg==5
xkit==0.0.0
zipp==1.0.0
